requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
agno==1.7.5
openai==1.58.1
firecrawl-py==2.16.2